Larger items shouldn't require contiguous write.
When they're small, it's cheaper to allocate a single buffer and make
a single Write() call.  When they get larger, it's cheaper to send the
header and then the other parts separately.  A good middle might be to
send the extras and keys and body separately depending on the size of
the body.