Consider memcached status != 0 as an error.
I kind of thought this happened already, but it's clear that it
didn't.  I struggled to decide whether this is good or not.  net/http,
for example, only errors if it couldn't properly *speak* http, but
makes no judgment as to the response codes.  I think this is less
right here as there's one right answer for any given request and the
function UnwrapMemcachedError was added for the minority cases where
it matters.