Consider EINVAL fatal
This really isn't generally a fatal error, but the error does show up in
a situation where protocol error causes the stream to fall out of sync.
The server will hang up, but the client won't know it yet.

This generally happens with a key that's too long for the server.

e.g. this test sets a value in a short key, then a long key, then tries
to read the value out of the short key again using the couchbase API.

before:

    len(longString) == 260
    Return:  MCResponse status=EINVAL, opcode=SET, opaque=0, msg: Invalid arguments
    EOF

after:

    len(longString) == 260
    Return:  MCResponse status=EINVAL, opcode=SET, opaque=0, msg: Invalid arguments
    x is "test"

cbugg: close bug-304