Return the most up-to-date CAS on success.
Normally CAS isn't checked in the success cases, except when
waiting on data to be persisted. Previously the library failed
to update the CAS value on success causing WriteUpdate calls with
the Index/Persist flags set to always fail with ErrOverwritten.

These were false errors, as the library was always using the
pre-modification CAS value as input into the observe function
instead of the post-modification CAS value.

cbugg: close bug-849